Moved api out of server package
